+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Latest telemetry data for given device & type
 - Assignment and unassignment policy to device
 - Device policy status for given device & policy & item type or list for given device by filter
+- Array `supportedResolutions` into Device entity
 
 ## [1.7.1] - 2021-12-17
 ### Fixed